[refactor/ABOUT-52] 생각 아이템 삭제 버튼 구현

+++ app/src/main/java/com/w36495/about/ui/fragment/ThinkListFragment.kt
@@ -2,7 +2,6 @@ package com.w36495.about.ui.fragment
 
 import android.content.Context
 import android.content.Intent
-import android.graphics.Point
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
@@ -17,19 +16,18 @@ import androidx.fragment.app.commit
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
-import com.google.android.material.appbar.MaterialToolbar
-import com.w36495.about.ui.ItemSwipeHelper
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
+import com.w36495.about.ui.listener.ThinkItemTouchHelper
 import com.w36495.about.R
 import com.w36495.about.ui.adapter.ThinkListAdapter
 import com.w36495.about.domain.entity.Think
-import com.w36495.about.contract.ThinkContract
 import com.w36495.about.contract.ThinkListContract
 import com.w36495.about.data.ThinkUiState
 import com.w36495.about.ui.presenter.ThinkListPresenter
 import com.w36495.about.domain.entity.Topic
 import com.w36495.about.data.local.AppDatabase
 import com.w36495.about.data.repository.ThinkRepositoryImpl
+import com.w36495.about.databinding.FragmentThinkListBinding
 import com.w36495.about.ui.dialog.ThinkDialogActivity
 import com.w36495.about.ui.listener.ThinkListItemClickListener
 import com.w36495.about.ui.listener.ThinkSwipeListener
@@ -38,6 +36,9 @@ import kotlinx.coroutines.launch
 
 class ThinkListFragment(private val topic: Topic) : Fragment(),
     ThinkSwipeListener, ThinkListItemClickListener, ThinkListContract.View {
+
+    private var _binding: FragmentThinkListBinding? = null
+    private val binding: FragmentThinkListBinding get() = _binding!!
 
     private lateinit var thinkListContext: Context
     private lateinit var thinkListView: View
@@ -53,10 +54,8 @@ class ThinkListFragment(private val topic: Topic) : Fragment(),
 
     private lateinit var getResultThink: ActivityResultLauncher<Intent>
 
-    private lateinit var toolbar: MaterialToolbar
-    private lateinit var recyclerView: RecyclerView
     private lateinit var thinkListAdapter: ThinkListAdapter
-    private lateinit var itemSwipeHelper: ItemSwipeHelper
+    private lateinit var thinkItemTouchHelper: ThinkItemTouchHelper
 
     private var database: AppDatabase? = null
     private lateinit var presenter: ThinkListContract.Presenter
@@ -64,14 +63,13 @@ class ThinkListFragment(private val topic: Topic) : Fragment(),
     private var position: Int? = null
     private var currentThink: Think? = null
 
-    private val size = Point()
-
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_think_list, container, false)
+        _binding = FragmentThinkListBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -81,21 +79,26 @@ class ThinkListFragment(private val topic: Topic) : Fragment(),
 
         database = AppDatabase.getInstance(view.context)
 
-        toolbar = view.findViewById(R.id.think_list_toolbar)
-        recyclerView = view.findViewById(R.id.think_list_recyclerview)
         presenter = ThinkListPresenter(ThinkRepositoryImpl(database!!.thinkDao()), this)
 
-        toolbar.title = topic.topic
+        binding.thinkListToolbar.title = topic.topic
 
         thinkListAdapter = ThinkListAdapter()
         thinkListAdapter.setClickListener(this, this)
 
-        itemSwipeHelper = ItemSwipeHelper(thinkListAdapter)
-        val itemTouchHelper = ItemTouchHelper(itemSwipeHelper)
-        itemTouchHelper.attachToRecyclerView(recyclerView)
+        thinkItemTouchHelper = ThinkItemTouchHelper()
+        val itemTouchHelper = ItemTouchHelper(thinkItemTouchHelper)
 
-        recyclerView.adapter = thinkListAdapter
-        recyclerView.layoutManager = LinearLayoutManager(view.context)
+        binding.thinkListRecyclerview.apply {
+            adapter = thinkListAdapter
+            layoutManager = LinearLayoutManager(view.context)
+
+            itemTouchHelper.attachToRecyclerView(this)
+            setOnTouchListener { _, _ ->
+                thinkItemTouchHelper.removePreviousSwipe(this)
+                false
+            }
+        }
 
         getResultThink = registerForActivityResult(
             ActivityResultContracts.StartActivityForResult()
@@ -123,7 +126,7 @@ class ThinkListFragment(private val topic: Topic) : Fragment(),
             }
         }
 
-        toolbar.setOnMenuItemClickListener { menu ->
+        binding.thinkListToolbar.setOnMenuItemClickListener { menu ->
             when (menu.itemId) {
                 R.id.main_add -> {
                     val moveThinkAddIntent =
@@ -136,7 +139,7 @@ class ThinkListFragment(private val topic: Topic) : Fragment(),
             }
         }
 
-        toolbar.setNavigationOnClickListener {
+        binding.thinkListToolbar.setNavigationOnClickListener {
             parentFragmentManager.popBackStack()
         }
 
@@ -167,7 +170,18 @@ class ThinkListFragment(private val topic: Topic) : Fragment(),
     }
 
     override fun onThinkSwiped(thinkId: Long) {
-        presenter.deleteThinkById(thinkId)
+        thinkItemTouchHelper.removeSwipeAfterDelete(binding.thinkListRecyclerview)
+
+        MaterialAlertDialogBuilder(thinkListContext, R.style.AboutTheme_AlertDialog)
+            .setTitle(resources.getString(R.string.dialog_title_think_delete))
+            .setMessage(resources.getString(R.string.dialog_message_think_delete))
+            .setNeutralButton(resources.getString(R.string.dialog_btn_cancel)) { dialog, _ ->
+                dialog.dismiss()
+            }
+            .setPositiveButton(resources.getString(R.string.dialog_btn_delete)) { dialog, _ ->
+                presenter.deleteThinkById(thinkId)
+            }
+            .show()
     }
 
     override fun onThinkListItemClicked(position: Int, think: Think) {
@@ -181,13 +195,6 @@ class ThinkListFragment(private val topic: Topic) : Fragment(),
 
     override fun showThink(think: Think) {
         currentThink = think
-//        position?.let {
-//            val moveThinkUpdateIntent = Intent(thinkListContext, ThinkDialogActivity::class.java)
-//            moveThinkUpdateIntent.putExtra("tag", DIALOG_UPDATE_TAG)
-//            moveThinkUpdateIntent.putExtra("position", it)
-//            moveThinkUpdateIntent.putExtra("think", think.text)
-//            getResultThink.launch(moveThinkUpdateIntent)
-//        }
 
         position?.let {
             parentFragmentManager.commit {
@@ -212,5 +219,10 @@ class ThinkListFragment(private val topic: Topic) : Fragment(),
         handler.postDelayed(Runnable {
             Toast.makeText(thinkListContext, message, Toast.LENGTH_SHORT).show()
         }, 0)
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
     }
 }
